refactor - follow a better directory structure based on https://github.com/slack-samples/bolt-ts-starter-template - update some types so that type is inferred for args to listeners - add clients to context for all requests instead of just the ones with messages middleware. move them to a different directory
