+++ src/lib/help.ts
@@ -1,41 +1,4 @@
-import { Context, SayFn } from "@slack/bolt";
-import { KnownEventFromType } from "@slack/bolt";
 import config from "./config";
-
-const helpRegex = /help\s*(all)?/i;
-
-async function help({
-  message,
-  context,
-  say,
-}: {
-  message: KnownEventFromType<"message">;
-  context: Context;
-  say: SayFn;
-}) {
-  if (message.subtype !== undefined && message.subtype !== "bot_message") {
-    return;
-  }
-
-  const isAll = isHelpAll(message.text ?? "");
-  let helpText;
-  if (isAll) {
-    helpText = helpAll(context.botUserId);
-  } else if (context.channel.name === config.FLARES_CHANNEL_NAME) {
-    helpText = helpFlaresChannel(context.botUserId);
-  } else if (context.channel.name.startsWith(config.FLARE_CHANNEL_PREFIX)) {
-    helpText = helpFlareChannel(context.botUserId);
-  } else {
-    helpText = helpAll(context.botUserId);
-  }
-  await say({
-    text: helpText,
-  });
-}
-
-function isHelpAll(text: string) {
-  return text.match(helpRegex)?.[1] === "all";
-}
 
 function helpFlaresChannel(botUserId: string | undefined) {
   return `
@@ -64,4 +27,4 @@ ${helpFlaresChannel(botUserId)}${helpFlareChannel(botUserId)}
 `;
 }
 
-export { helpRegex, help, helpAll, helpFlaresChannel };
+export { helpAll, helpFlaresChannel, helpFlareChannel };
