+++ src/middleware/index.ts
@@ -1,0 +1,8 @@
+import type { AnyMiddlewareArgs, App, Middleware } from "@slack/bolt";
+import { messageMiddleware } from "./message";
+
+const register = (app: App) => {
+  app.use(messageMiddleware as Middleware<AnyMiddlewareArgs>);
+};
+
+export default { register };
